requirements.txt
streamlit>=1.33.0
requests>=2.31.0
beautifulsoup4>=4.12.3
newspaper3k>=0.2.8
lxml_html_clean>=0.1.0
transformers>=4.41.1
torch>=2.3.0
nltk
sumy
rake-nltk
keybert
sentence-transformers
huggingface_hub[hf_xet]
